+++ Desktop/DrawPilling/src/Scripts/NewDraw.cpp
@@ -154,7 +154,7 @@ void NewDraw::InitBrush(RenderData& data, float& radius, std::string texture)
 	}
 }
 
-void NewDraw::BrushToPosition(GLFWwindow* window, RenderData& cursor, float& radius, float* aspect, float* offset, float* scale, float* position, int debug) {
+void NewDraw::BrushToPosition(GLFWwindow* window, RenderData& cursor, float& radius, float* aspect, float* offset, float* scale, double* position, int debug) {
 	float positions[16];
 	float yMult = aspect[0] / aspect[1];
 	fillPositions(positions, radius / aspect[0] * scale[0], radius * yMult / aspect[0] * scale[1], (position[0] - offset[0]) / aspect[0], (position[1] - offset[1]) / aspect[1]);
